+++ libs/blocks/text/text.css
@@ -136,54 +136,6 @@
   font-style: italic;
 }
 
-/* link-farm */
-.link-farm.text-block {
-  padding: var(--spacing-xl) 0;
-}
-
-.link-farm.text-block .foreground {
-  display: grid;
-  grid-template-columns: minmax(0, 1fr);
-  gap: var(--spacing-xl) var(--spacing-m);
-}
-
-.link-farm.text-block .foreground:first-child {
-  grid-template-columns: auto;
-}
-
-.link-farm.text-block h3 {
-  margin: 0 0 var(--spacing-l);
-  font-size: var(--type-heading-xs-size);
-  line-height: var(--type-heading-xs-lh);
-}
-
-.link-farm.text-block .no-heading {
-  margin: 0;
-}
-
-.link-farm.text-block h2 {
-  margin: 0 0 var(--spacing-m);
-}
-
-.link-farm.text-block p {
-  font-size: var(--type-body-s-size);
-  margin: 0 0 var(--spacing-s);
-  line-height: var(--type-body-s-lh);
-}
-
-.link-farm.text-block p:last-child {
-  margin: 0;
-}
-
-.link-farm.text-block a {
-  color: inherit;
-  text-decoration: underline;
-}
-
-.link-farm.text-block a:hover {
-  text-decoration: double underline;
-}
-
 .section[class*='grid-width-'] .text-block .foreground {
   max-width: unset;
 }
@@ -239,31 +191,8 @@
   display: none;
 }
 
-
-/* mobile */
-@media screen and (max-width: 600px) {
-  .link-farm.text-block .foreground {
-    row-gap: 0;
-  }
-
-  .link-farm.text-block .has-heading {
-    margin-top: var(--spacing-s);
-  }
-
-  .link-farm.text-block .foreground div:last-child p:last-child{
-    margin-bottom: 0;
-  }
-
-  .link-farm.text-block p:last-child{
-    margin-bottom: var(--spacing-s);
-  }
-}
-
 /* Tablet up */
 @media screen and (min-width: 600px) {
-  .link-farm.text-block h2 {
-    margin: 0 0 var(--spacing-xl-static);
-  }
 
   .inset.text-block .foreground::before {
     display: none;
@@ -277,40 +206,6 @@
     padding-left: var(--spacing-l);
   }
 
-  .link-farm.text-block .foreground {
-    grid-template-columns: repeat(2, minmax(0, 1fr));
-  }
-
-  .link-farm.text-block .no-heading {
-    display: none;
-  }
-
-  .link-farm.text-block .hspace .no-heading {
-    display: block;
-  }
-
-  .link-farm.text-block .no-heading {
-    height: var(--type-heading-m-size);
-    margin: 0 0 var(--spacing-l);
-  }
-
-  .link-farm.text-block .foreground {
-    row-gap: var(--spacing-s);
-  }
-
-   .link-farm.text-block .gap-xl {
-    row-gap: var(--spacing-xl);
-   }
-
-
-  .link-farm.text-block .foreground div p:last-child{
-    margin-bottom: 0;
-  }
-
-  .link-farm.text-block p:last-child{
-    margin-bottom: var(--spacing-s);
-  }
-  
   .text-block .foreground > .tablet-up,
   .text-block .foreground > .tablet-up.desktop-up {
     display: block;
@@ -339,14 +234,6 @@
 
 /* Desktop up */
 @media screen and (min-width: 1200px) {
-  .link-farm.text-block .foreground {
-    grid-template-columns: repeat(4, minmax(0, 1fr));
-  }
-
-  .link-farm.text-block .no-heading {
-    display: block;
-  }
-  
   .text-block .foreground > .desktop-up,
   .text-block .foreground > .tablet-up.desktop-up {
     display: block;
