+++ libs/blocks/text/text.js
@@ -1,5 +1,5 @@
 import { decorateBlockBg, decorateBlockText, getBlockSize, decorateTextOverrides } from '../../utils/decorate.js';
-import { createTag } from '../../utils/utils.js';
+import decorateLinkFarms from '../link-farms/link-farms.js';
 
 // size: [heading, body, ...detail]
 const blockTypeSizes = {
@@ -81,27 +81,8 @@ export default function init(el) {
     const elAction = el.querySelector('.action-area');
     if (elAction) elAction.classList.add('body-s');
   }
-  if (el.classList.contains('link-farm')) {
-    const foregroundDiv = el.querySelectorAll('.foreground')[1];
-    const count = foregroundDiv.querySelectorAll('h3').length;
-    foregroundDiv.querySelectorAll('div').forEach((divElem, index) => {
-      const h3 = divElem.querySelector('h3');
-      if (count) {
-        if (h3) {
-          const sibling = index % 2 === 0
-            ? divElem.nextElementSibling
-            : divElem.previousElementSibling;
-          sibling?.classList.add('hspace');
-          if (index > 0) divElem.classList.add('has-heading');
-          if (index > 1) foregroundDiv.classList.add('gap-xl');
-        } else {
-          const headingElem = createTag('h3', { class: 'no-heading' });
-          divElem.insertBefore(headingElem, divElem.firstChild);
-        }
-      }
-    });
-  }
+  if (el.classList.contains('link-farm')) decorateLinkFarms(el);
   el.classList.add(...helperClasses);
   decorateTextOverrides(el);
-  decorateMultiViewport(el);
+  if (!el.classList.contains('link-farm')) decorateMultiViewport(el);
 }
